Split range module
